wildlens_api/requirements.txt
numpy<2.0.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.6
torch==2.2.2
torchvision==0.17.2
Pillow==10.2.0
requests==2.31.0
email-validator==2.1.1
cloudinary
